Move QgsOgrProvider::addSubLayerDetailsToSubLayerList to QgsOgrProviderUtils (no code changes, just movement)
