+++ src/core/providers/ogr/qgsogrprovider.h
@@ -198,8 +198,6 @@ class QgsOgrProvider final: public QgsVectorDataProvider
     //! Does the real job of settings the subset string and adds an argument to disable update capabilities
     bool _setSubsetString( const QString &theSQL, bool updateFeatureCount = true, bool updateCapabilities = true, bool hasExistingRef = true );
 
-    void addSubLayerDetailsToSubLayerList( int i, QgsOgrLayer *layer, bool withFeatureCount ) const;
-
     QStringList _subLayers( bool withFeatureCount ) const;
 
     QgsFields mAttributeFields;
